Quick fix

+++ src/fr/hugob147/endorialobby/coins/Coins.java
@@ -9,7 +9,7 @@ import java.sql.SQLException;
 
 public class Coins
 {
-	public void createAccount(Player player, long coins)
+	public void createAccount(Player player, Double coins)
 	{
 		try
 		{
@@ -21,7 +21,7 @@ public class Coins
 				sts.close();
 				sts = MySQL.getConnection().prepareStatement("INSERT INTO coins (player_uuid, coins) VALUES (?, ?)");
 				sts.setString(1, player.getUniqueId().toString());
-				sts.setLong(2, coins);
+				sts.setDouble(2, coins);
 				sts.executeUpdate();
 			}
 		} catch (SQLException e)
@@ -30,7 +30,7 @@ public class Coins
 		}
 	}
 
-	public long getCoins(Player player)
+	public Double getCoins(Player player)
 	{
 		try
 		{
@@ -39,16 +39,16 @@ public class Coins
 			ResultSet rs = sts.executeQuery();
 			if (rs.next())
 			{
-				return rs.getLong("coins");
+				return rs.getDouble("coins");
 			}
 		} catch (SQLException e)
 		{
 			e.printStackTrace();
 		}
-		return 0L;
+		return 0D;
 	}
 
-	public void addCoins(Player player, long coins)
+	public void addCoins(Player player, Double coins)
 	{
 		if (coins < 1L)
 		{
@@ -61,10 +61,10 @@ public class Coins
 			ResultSet rs = sts.executeQuery();
 			if (rs.next())
 			{
-				long money = rs.getLong("coins");
+				Double money = rs.getDouble("coins");
 				sts.close();
 				sts = MySQL.getConnection().prepareStatement("UPDATE coins SET coins=? WHERE player_uuid=?");
-				sts.setLong(1, coins + money);
+				sts.setDouble(1, coins + money);
 				sts.setString(2, player.getUniqueId().toString());
 				sts.executeUpdate();
 			}
@@ -74,7 +74,7 @@ public class Coins
 		}
 	}
 
-	public void removeCoins(Player player, long coins)
+	public void removeCoins(Player player, Double coins)
 	{
 		if (coins < 1L)
 		{
@@ -87,14 +87,14 @@ public class Coins
 			ResultSet rs = sts.executeQuery();
 			if (rs.next())
 			{
-				long money = rs.getLong("coins");
+				Double money = rs.getDouble("coins");
 				sts.close();
 				if (money - coins < 0L)
 				{
 					coins = money;
 				}
 				sts = MySQL.getConnection().prepareStatement("UPDATE coins SET coins=? WHERE player_uuid=?");
-				sts.setLong(1, money - coins);
+				sts.setDouble(1, money - coins);
 				sts.setString(2, player.getUniqueId().toString());
 				sts.executeUpdate();
 			}
